refactor: move paths to /posthog
