+++ ee/clickhouse/queries/test/test_paths.py
@@ -10,7 +10,7 @@ from freezegun import freeze_time
 from ee.clickhouse.models.group import create_group
 from ee.clickhouse.models.session_recording_event import create_session_recording_event
 from ee.clickhouse.queries.paths import ClickhousePaths, ClickhousePathsActors
-from ee.clickhouse.queries.paths.path_event_query import PathEventQuery
+from ee.clickhouse.queries.paths.paths_event_query import PathEventQuery
 from ee.clickhouse.util import ClickhouseTestMixin, snapshot_clickhouse_queries
 from posthog.constants import (
     FUNNEL_PATH_AFTER_STEP,
