+++ ee/clickhouse/views/insights.py
@@ -6,6 +6,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 
 from ee.clickhouse.queries.funnels.funnel_correlation import FunnelCorrelation
+from ee.clickhouse.queries.paths import ClickhousePaths
 from ee.clickhouse.queries.retention import ClickhouseRetention
 from ee.clickhouse.queries.stickiness import ClickhouseStickiness
 from posthog.api.insight import InsightViewSet
@@ -32,6 +33,7 @@ class ClickhouseInsightsViewSet(InsightViewSet):
 
     retention_query_class = ClickhouseRetention
     stickiness_query_class = ClickhouseStickiness
+    paths_query_class = ClickhousePaths
 
     # ******************************************
     # /projects/:id/insights/funnel/correlation
