+++ posthog/queries/paths/__init__.py
@@ -1,2 +1,3 @@
+# flake8: noqa
 from .paths import *
 from .paths_actors import *
